moved assets into respective assets folder
Moved images from the root of src into an assets folder. Updated all imports accordingly.
